#49 Deleted search service
