+++ app/Services/Trip/TripServiceInterface.php
@@ -12,6 +12,7 @@ use Illuminate\Contracts\Pagination\LengthAwarePaginator;
 interface TripServiceInterface
 {
     public function createTrip(array $data, User $user): void;
+    public function searchTrips(string $searchRequest, ?string $perPage): LengthAwarePaginator;
     public function getTrips(QueryStringData $data, User $user, ?string $perPage): LengthAwarePaginator;
     public function getUserTrips(User $user, User $loggedUser, ?string $perPage): LengthAwarePaginator;
     public function updateTrip(Trip $trip, array $data): void;
